move dashboard utils
